Refactor the backfill to call a worker instead
